+++ core/src/main/java/io/snabble/sdk/events/Events.kt
@@ -5,7 +5,7 @@ import android.os.Looper
 import android.os.SystemClock
 import io.snabble.sdk.Project
 import io.snabble.sdk.Shop
-import io.snabble.sdk.ShoppingCart
+import io.snabble.sdk.shoppingcart.ShoppingCart
 import io.snabble.sdk.Snabble
 import io.snabble.sdk.codes.ScannedCode
 import io.snabble.sdk.events.data.Event
