+++ core/src/main/java/io/snabble/sdk/shoppingcart/data/item/CouponItem.kt
@@ -1,4 +1,4 @@
-package io.snabble.sdk.shoppingcart.data
+package io.snabble.sdk.shoppingcart.data.item
 
 import io.snabble.sdk.codes.ScannedCode
 import io.snabble.sdk.coupons.Coupon
