+++ core/src/main/java/io/snabble/sdk/shoppingcart/data/item/ItemType.kt
@@ -1,4 +1,4 @@
-package io.snabble.sdk.shoppingcart.data
+package io.snabble.sdk.shoppingcart.data.item
 
 /**
  * Enum describing the type of item
