+++ core/src/test/java/io/snabble/sdk/checkout/MockCheckoutApi.kt
@@ -3,7 +3,7 @@ package io.snabble.sdk.checkout
 import com.google.gson.JsonObject
 import io.snabble.sdk.PaymentMethod
 import io.snabble.sdk.Project
-import io.snabble.sdk.ShoppingCart
+import io.snabble.sdk.shoppingcart.ShoppingCart
 import io.snabble.sdk.merge
 import io.snabble.sdk.payment.PaymentCredentials
 import io.snabble.sdk.utils.GsonHolder
