+++ ui/src/main/java/io/snabble/sdk/ui/cart/ShoppingCartItemViewHolder.kt
@@ -20,7 +20,7 @@ import com.squareup.picasso.Picasso
 import io.snabble.accessibility.accessibility
 import io.snabble.accessibility.orderViewsForAccessibility
 import io.snabble.sdk.Product
-import io.snabble.sdk.ShoppingCart
+import io.snabble.sdk.shoppingcart.ShoppingCart
 import io.snabble.sdk.Snabble
 import io.snabble.sdk.ui.R
 import io.snabble.sdk.ui.cart.ShoppingCartView.ProductRow
