+++ ui/src/main/java/io/snabble/sdk/ui/cart/ShoppingCartView.java
@@ -32,11 +32,11 @@ import java.util.List;
 import io.snabble.sdk.PriceFormatter;
 import io.snabble.sdk.Product;
 import io.snabble.sdk.Project;
-import io.snabble.sdk.ShoppingCart;
+import io.snabble.sdk.shoppingcart.ShoppingCart;
 import io.snabble.sdk.Snabble;
 import io.snabble.sdk.Unit;
 import io.snabble.sdk.ViolationNotification;
-import io.snabble.sdk.shoppingcart.data.ItemType;
+import io.snabble.sdk.shoppingcart.data.item.ItemType;
 import io.snabble.sdk.shoppingcart.data.listener.ShoppingCartListener;
 import io.snabble.sdk.shoppingcart.data.listener.SimpleShoppingCartListener;
 import io.snabble.sdk.ui.GestureHandler;
