+++ ui/src/main/java/io/snabble/sdk/ui/cart/UndoHelper.java
@@ -1,6 +1,6 @@
 package io.snabble.sdk.ui.cart;
 
-import io.snabble.sdk.ShoppingCart;
+import io.snabble.sdk.shoppingcart.ShoppingCart;
 
 public interface UndoHelper {
     void removeAndShowUndoSnackbar(final int adapterPosition, final ShoppingCart.Item item);
